Fix OpenRD configuration (Closes: #793214)
In a similar fashion to the SheevaPlug the upstream FTDI2232 support
changes broke OpenRD support. The fix requires a change to the channel
used, the USB device description updated so the adaptor is found and
the addition of the JTAG speed in the board configuration. Patch has
been forwarded upstream.